removed unnecessary float convert

+++ internal/memtrack/http/client/client.go
@@ -26,7 +26,7 @@ type Client struct {
 func (c Client) SendCounter(metric metrics.Metricable, mapMetrics map[string]interface{}) {
 	log.Printf("sending to host: %s", c.Host)
 	for k, v := range mapMetrics {
-		val := float64(v.(float64))
+		val := v.(float64)
 		del := int64(val)
 		toMarsal := metrics.Metrics{
 			ID:    k,
@@ -59,7 +59,7 @@ func (c Client) SendCounter(metric metrics.Metricable, mapMetrics map[string]int
 func (c Client) SendGauges(metric metrics.Metricable, mapMetrics map[string]interface{}) {
 	log.Printf("sending to host: %s", c.Host)
 	for k, v := range mapMetrics {
-		val := float64(v.(float64))
+		val := v.(float64)
 		toMarsal := metrics.Metrics{
 			ID:    k,
 			MType: metric.String(),
